+++ src/util/CMakeLists.txt
@@ -10,8 +10,6 @@ nebula_add_library(
     SchemaUtil.cpp
     IndexUtil.cpp
     ZoneUtil.cpp
-    
-    GroupUtil.cpp
     ToJson.cpp
 )
