--- src/util/GroupUtil.cpp
@@ -1,14 +1,0 @@
-/* Copyright (c) 2020 vesoft inc. All rights reserved.
- *
- * This source code is licensed under Apache 2.0 License,
- * attached with Common Clause Condition 1.0, found in the LICENSES directory.
- */
-
-#include "util/GroupUtil.h"
-
-namespace nebula {
-namespace graph {
-
-
-}   // namespace graph
-}   // namespace nebula
