--- src/util/GroupUtil.h
@@ -1,24 +1,0 @@
-/* Copyright (c) 2020 vesoft inc. All rights reserved.
- *
- * This source code is licensed under Apache 2.0 License,
- * attached with Common Clause Condition 1.0, found in the LICENSES directory.
- */
-
-#ifndef UTIL_GROUPUTIL_H_
-#define UTIL_GROUPUTIL_H_
-
-#include "common/base/Base.h"
-#include "common/base/StatusOr.h"
-#include "parser/MaintainSentences.h"
-
-namespace nebula {
-namespace graph {
-
-class GroupUtil final {
-public:
-    GroupUtil() = delete;
-};
-
-}  // namespace graph
-}  // namespace nebula
-#endif  // UTIL_GROUPUTIL_H_
